services/orchestrator/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==6.0.0
prometheus-client==0.20.0
